requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
constantly==23.10.4
cryptography==42.0.6
daphne==4.1.2
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
dnspython==2.6.1
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
ifaddr==0.2.0
incremental==22.10.0
joblib==1.4.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
py_eureka_client==0.11.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
zope.interface==6.3
